requirements.txt
# AI Workshop Requirements - Compatible Versions
# Updated for Python 3.8+ compatibility with PyTorch 2.6.x

# Core scientific computing
numpy==1.26.4
pandas==2.2.0

# Deep learning frameworks
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
tensorflow==2.18.0

# Transformers and NLP
transformers==4.46.0
sentence-transformers==3.3.0
datasets==3.1.0
accelerate==1.1.0
tokenizers==0.20.1

# Computer vision
opencv-python==4.10.0.84
pillow==10.4.0
ultralytics==8.3.0

# Visualization and plotting
matplotlib==3.9.2
seaborn==0.13.2

# Machine learning
scikit-learn==1.5.2
faiss-cpu==1.8.0

# Web and API
requests==2.32.3
beautifulsoup4==4.12.3
gradio==5.5.0

# Jupyter and notebooks
jupyter==1.1.1
ipywidgets==8.1.5
notebook==7.2.2

# Utility libraries
tqdm==4.66.6
psutil==6.1.0

# Testing
pytest==8.3.3